requirements.txt
# Telegram Bot
python-telegram-bot==20.8
python-telegram-bot[job-queue]==20.8

# Telegram Client for reading channels
telethon==1.34.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.1

# Scheduler
APScheduler==3.10.4

# Database (SQLite is built-in, but we need this for async support)
aiosqlite==0.19.0
